backend/requirements.txt
fastapi
uvicorn
numpy<2.0.0
pandas>=1.5.0
scikit-learn>=1.2.0
scikit-learn-extra
joblib
matplotlib
seaborn
openpyxl
chardet
tpot
flaml
h2o
imbalanced-learn
huggingface-hub>=0.20.0
transformers>=4.36.0
sentence-transformers
python-dotenv
torch --index-url https://download.pytorch.org/whl/cpu
google-generativeai